net: moved and minimized the netdev interface
