--- drivers/netdev/802154/Makefile
@@ -1,5 +1,0 @@
-MODULE := netdev_802154
-
-INCLUDES += -I$(RIOTBASE)/drivers/include
-
-include $(RIOTBASE)/Makefile.base
