--- drivers/netdev/base/Makefile
@@ -1,5 +1,0 @@
-MODULE := netdev_base
-
-INCLUDES += -I$(RIOTBASE)/drivers/include
-
-include $(RIOTBASE)/Makefile.base
